Finally found that bloody combination of two bugs
- there was a mistake in the boundary conditions where it moved the hot
area towards x+/y+ direction but it looked almost symmetrically on finer
grids --> This is a strong case against procedural boundary settings on
the individual grids and in favor of procedural boundary settings on the
finest grid plus coarsening it to the other grids!
- also changed smoothing step to alternate between iterating upwards and
downwards in all dimensions. Otherwise one corner converges slightly
quicker due to the Gauss-Seidel-like iteration. This is removed by
successive smoothing steps but it intermediately destroys symmetry and
is therefore ugly.