lagging async global residual evaluation
reorganized residual computation such that the global residual is
collected in an async fashion. the resulting global residual is used
only in the following iteration to determine if the loop is stopped or
continued. This leads to one additional iteration but avoids waiting for
the allreduce in every single iteration.