fixes and test
* Generously added more barriers, because there where stray value
errors. Some of those barriers are not needed I guess but some really
are. Come back to this to keep only required barriers.
* Added a check for symmetry in the final solution on the finest grid.
With proper (i.e. symetric) boundary conditions this tells if there are
value errores even without the CSV output and the visual check via
paraview.
* Removed the barriers in the writetocsv routine. The presence or
absence of this must not influence the logic of the barriers.
* Extra parameters in MiniMon are only needed at the latter calls per
region, not at the first one.
* Track smoothen function with residual separately in MiniMon